// src/types/exhibition.type.ts
/* 
{
    "addr1": "서울특별시 마포구 월드컵북로 31 아만티호텔 서울",
    "addr2": "",
    "areacode": "1",
    "booktour": "",
    "cat1": "B02",
    "cat2": "B0201",
    "cat3": "B02010100",
    "contentid": "2917245",
    "contenttypeid": "32",
    "createdtime": "20221124101358",
    "firstimage": "http://tong.visitkorea.or.kr/cms/resource/87/2916587_image2_1.jpg",
    "firstimage2": "http://tong.visitkorea.or.kr/cms/resource/87/2916587_image3_1.jpg",
    "cpyrhtDivCd": "Type3",
    "mapx": "126.9187184009",
    "mapy": "37.5571903727",
    "mlevel": "6",
    "modifiedtime": "20240216173551",
    "sigungucode": "13",
    "tel": "",
    "title": "아만티호텔서울",
    "zipcode": "04001"
}
*/
export interface ExhibitionType {
  addr1: string;
  addr2: string;
  areacode: string;
  booktour: string;
  cat1: string;
  cat2: string;
  cat3: string;
  contentid: string;
  contenttypeid: string;
  createdtime: string;
  firstimage: string;
  firstimage2: string;
  cpyrhtDivCd: string;
  mapx: string;
  mapy: string;
  mlevel: string;
  modifiedtime: string;
  sigungucode: string;
  tel: string;
  title: string;
  zipcode: string;
}